requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==25.1.0
click==8.2.1
coverage==7.9.1
decorator==5.2.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
